:bug: Ensure the openzaak-private-media volume permissions are correct
Bugreport https://github.com/open-zaak/open-zaak/issues/481 revealed
that the nginx user has no permissions to read the openzaak-private-media
volume. This is because the Docker daemon runs as root user, and the
volume created is owned by root. Uploads work, because uwsgi runs as
root (to be fixed soon-ish), thus it can write to the directory.

Nginx however, runs as a non-privileged user (a good thing), but
this means that it has no read permissions on the volume, which it needs
to serve the files of course.

The patch ensures a group is created for openzaak, which contains the
openzaak user and nginx. The volume directory permissions are managed
so that the openzaak user can write (irrelevant for now, since the
application runs as root still), and the group can read the volume.
Effectively this adds the ability for nginx to serve the files from
the volume. Any other system users are not even allowed to read the
files.