:sparkles: Added PostgreSQL 10 role for SUSE Linux
Added a role to install a PostgreSQL (10) cluster with Postgis support
on SUSE Linux. Tested with OpenSUSE Leap 15.1, should also work with
SLES 15SP1.

Postgis support is iffy on SUSE - there is no official package for
PostgreSQL 11, but there is for PostgreSQL 10. Make sure to specify
the 'postgresql_version' variable on subsequent roles that require this
version number, as Open Zaak/Open Notificaties defaults to PG 11.

This contribution was sparked by the municipality of Delft (https://delft.nl).