:bug: Fixes open-zaak/open-zaak#922 -- set nginx location subpath correctly
If Open Zaak/Open Notificaties is hosted on a subpath, this should be
reflected in the nginx configuration that routes traffic to the upstream
containers. URLs will typically contain this subpath, so the location
match must typically include the prefix.

The roles derive the value automatically from the foo_subpath variable
under normal circumstances.