:bug: podSecurityContext.runAsUser|runAsGroup is too strict
The idea of making this strict is a good one, but too optimistic. Various
Kubernetes add-ons inject sidecars/init containers which run as a designated
container user ID which is not compatible with this security context.

Having the ability to specify this more strictly is still important,
which is why the default value is modified.